refactor: move modals to own components
