+++ src/features/profile/components/Person/Person.module.scss
@@ -27,12 +27,6 @@
 				font-weight: var(--text-weight-md);
 				letter-spacing: var(--text-letter-spacing);
 			}
-
-			.status {}
-
-			.pointer {
-				cursor: pointer;
-			}
 		}
 	}
 }
