+++ src/features/profile/components/Person/Person.tsx
@@ -1,5 +1,5 @@
-import {AvatarModalContent} from '@/features/profile/components/Person/AvatarModalContent/AvatarModalContent.tsx';
-import {StatusModalContent} from '@/features/profile/components/Person/StatusModalContent/StatusModalContent.tsx';
+import {AvatarControl} from "@/features/profile/components/AvatarControl/AvatarControl.tsx";
+import {StatusControl} from "@/features/profile/components/StatusControl/StatusControl.tsx";
 import anotherUserBcgWebp from '@/assets/images/profile-cover/another-user.webp';
 import {ExtraInfo} from "@/features/profile/components/ExtraInfo/ExtraInfo.tsx";
 import profileBcgWebp from '@/assets/images/profile-cover/profile-bcg.webp';
@@ -7,10 +7,7 @@ import anotherUserBcg from '@/assets/images/profile-cover/another-user.jpg';
 import cls from '@/features/profile/components/Person/Person.module.scss';
 import profileBcg from '@/assets/images/profile-cover/profile-bcg.jpeg';
 import {UserProfileType} from '@/features/profile/profile-reducer.ts';
-import {PropsWithChildren, FC, useState, useCallback} from 'react';
-import {Avatar} from '@/components/Avatar/Avatar.tsx';
-import {Modal} from '@/components/Modal/Modal.tsx';
-import classNames from 'classnames';
+import {PropsWithChildren, FC} from 'react';
 
 type PersonPropsType = PropsWithChildren & {
     profile: UserProfileType,
@@ -21,32 +18,6 @@ type PersonPropsType = PropsWithChildren & {
 export const Person: FC<PersonPropsType> = ({isMe, status, profile, children}) => {
     const {fullName, photos: {large: photo}} = profile;
 
-    const [photoIsOpen, setPhotoIsOpen] = useState(false);
-
-    const [statusEditMode, setStatusEditMode] = useState(false);
-
-    const openPhotoModal = useCallback(() => {
-        setPhotoIsOpen(true);
-    }, []);
-
-    const closePhotoModal = useCallback(() => {
-        setPhotoIsOpen(false);
-    }, []);
-
-    const openStatusModal = useCallback(() => {
-        if (isMe) setStatusEditMode(true);
-    }, []);
-
-    const closeStatusModal = useCallback(() => {
-        setStatusEditMode(false);
-    }, []);
-
-    const noStatusPlaceholder = isMe ? 'Enter status...' : ''
-
-    const classes = {
-        status: classNames(cls.status, {[cls.pointer]: isMe})
-    }
-
     return (
         <section>
             <picture>
@@ -55,25 +26,9 @@ export const Person: FC<PersonPropsType> = ({isMe, status, profile, children}) =
             </picture>
             <div className={cls.person}>
                 <div className={cls.info}>
-                    <Avatar size='12.5rem' photo={photo} border onClick={openPhotoModal} turnOffCursorPointer={!photo}/>
-                    {
-                        photo && <Modal onClose={closePhotoModal} opened={photoIsOpen}>
-                            <AvatarModalContent photo={photo} name={fullName}/>
-                        </Modal>
-                    }
-                    {
-                        isMe && <Modal onClose={closeStatusModal} opened={statusEditMode}>
-                            <StatusModalContent onClose={closeStatusModal}/>
-                        </Modal>
-                    }
+                    <AvatarControl photo={photo} fullName={fullName} />
                     <div className={cls.about}>
-                        <p className={classes.status} onClick={openStatusModal}>
-                            {
-                                status
-                                    ? status
-                                    : noStatusPlaceholder
-                            }
-                        </p>
+                        <StatusControl isMe={isMe} status={status} />
                         <h2 className={cls.name}>{fullName}</h2>
                     </div>
                 </div>
